Move source file; fix tests.
