+++ TypedPriorityQueue.ts
